Create question generator
The question generator has a set of categories and topics for each of
these categories.

The generator then chooses a random category and then random topics
within that category to be answer choices. One of those answer
choices are then deemed the correct answer. This answer choice is then
searched through the NASA Image API to get an image related to the
correct answer.